(DEPRECATED) 1.0.2
Added deprecation warning to readme. This is because KTERA is getting the official Baldera patch soon. When we get the patch on non-KR regions, this mod will be redundant.

It's a bit early, but I'd rather get people away from using the mod beforehand than have people deal with duplicated stuff after the patch hits.

This mod should still function, as I don't expect anything to break from v107 to whenever we get the Baldera patch.